# Implement caching for UTI and DateFormatters