--- a/NextcloudKit/NKModel.swift
+++ b/NextcloudKit/NKModel.swift
@@ -336,7 +336,6 @@ import SwiftyJSON
     @objc public var checksums = ""
     @objc public var creationDate: NSDate?
     @objc public var dataFingerprint = ""
-    @objc public var date = NSDate()
     @objc public var directory: Bool = false
     @objc public var downloadURL = ""
     @objc public var e2eEncrypted: Bool = false
@@ -380,6 +379,15 @@ import SwiftyJSON
     @objc public var urlBase = ""
     @objc public var user = ""
     @objc public var userId = ""
+
+    @objc public var dateString = ""
+    @objc public lazy var date: NSDate = {
+        if let date = NKCommon.shared.convertDate(dateString, format: "EEE, dd MMM y HH:mm:ss zzz") {
+            return date
+        } else {
+            return NSDate()
+        }
+    }()
 }
 
 @objc public class NKNotifications: NSObject {
@@ -1122,9 +1130,7 @@ class NKDataFileXML: NSObject {
             let propstat = element["d:propstat"][0]
                         
             if let getlastmodified = propstat["d:prop", "d:getlastmodified"].text {
-                if let date = NKCommon.shared.convertDate(getlastmodified, format: "EEE, dd MMM y HH:mm:ss zzz") {
-                    file.date = date
-                }
+                file.dateString = getlastmodified
             }
             
             if let creationtime = propstat["d:prop", "nc:creation_time"].double {